Split into editor and viewer folders
