+++ novelwriter/editor/autoreplace.py
