+++ novelwriter/editor/completer.py
