+++ novelwriter/editor/editfooter.py
