+++ novelwriter/editor/textblock.py
