+++ novelwriter/gui/itemdetails.py
