+++ novelwriter/gui/mainmenu.py
