+++ novelwriter/gui/search.py
