+++ novelwriter/gui/statusbar.py
