+++ tests/gui/test_outline.py
