+++ tests/gui/test_sidebar.py
