+++ tests/gui/test_statusbar.py
