+++ tests/viewer/test_viewerpanel.py
